GObject: Move closure handling into safehandle The explicit disconnect of a closure only happend if an object instance got disposed manually by a user. By moving the code into the the ObjectHandle the signals are disconnected in all cases.
