+++ src/Libs/GObject-2.0/Internal/ObjectHandle.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -7,6 +8,8 @@ namespace GObject.Internal;
 
 public class ObjectHandle : SafeHandle
 {
+    private readonly Dictionary<(uint, Delegate), (CULong, GObject.Closure)> _signalStore = new();
+
     public override bool IsInvalid => handle == IntPtr.Zero;
 
     public ObjectHandle(IntPtr handle, bool ownsHandle) : base(IntPtr.Zero, true)
@@ -22,15 +25,15 @@ public class ObjectHandle : SafeHandle
             // - Unowned GObjects need to be refed to bind them to this instance
             // - Unowned InitiallyUnowned floating objects need to be ref_sinked
             // - Unowned InitiallyUnowned non-floating objects need to be refed
-            // As ref_sink behaves like ref in case of non floating instances we use it for all 3 cases
+            // As ref_sink behaves like ref in the case of non-floating instances, we use it for all 3 cases
             Object.RefSink(handle);
         }
         else
         {
-            //In case we own the ref because the ownership was fully transfered to us we
+            //In case we own the ref because the ownership was fully transferred to us, we
             //do not need to ref the object at all.
 
-            Debug.Assert(!Internal.Object.IsFloating(handle), $"Handle {handle}: Owned floating references are not possible.");
+            Debug.Assert(!Object.IsFloating(handle), $"Handle {handle}: Owned floating references are not possible.");
         }
     }
 
@@ -41,9 +44,42 @@ public class ObjectHandle : SafeHandle
         InstanceCache.Add(handle, obj);
     }
 
+    internal void Connect(SignalDefinition signalDefinition, Delegate callback, GObject.Closure closure, bool after, string? detail)
+    {
+        var detailQuark = GLib.Functions.QuarkFromString(detail);
+        var handlerId = Functions.SignalConnectClosureById(DangerousGetHandle(), signalDefinition.Id, detailQuark, closure.Handle, after);
+
+        if (handlerId.Value == 0)
+            throw new Exception($"Could not connect to event {signalDefinition.ManagedName}");
+
+        _signalStore[(signalDefinition.Id, callback)] = (handlerId, closure);
+    }
+
+    internal void Disconnect(SignalDefinition signalDefinition, Delegate callback)
+    {
+        if (!_signalStore.TryGetValue((signalDefinition.Id, callback), out var tuple))
+            return;
+
+        Functions.SignalHandlerDisconnect(DangerousGetHandle(), tuple.Item1);
+        tuple.Item2.Dispose();
+        _signalStore.Remove((signalDefinition.Id, callback));
+    }
+
+    private void DisconnectSignals()
+    {
+        foreach (var item in _signalStore.Values)
+        {
+            Functions.SignalHandlerDisconnect(DangerousGetHandle(), item.Item1);
+            item.Item2.Dispose();
+        }
+
+        _signalStore.Clear();
+    }
+
     protected override bool ReleaseHandle()
     {
         RemoveMemoryPressure();
+        DisconnectSignals();
         InstanceCache.Remove(handle);
         return true;
     }
