+++ src/Libs/GObject-2.0/Public/Object.cs
@@ -19,7 +19,6 @@ public partial class Object : IDisposable
     public virtual void Dispose()
     {
         Debug.WriteLine($"Handle {Handle.DangerousGetHandle()}: Disposing object of type {GetType()}.");
-        DisposeClosures();
         Handle.Dispose();
     }
 }
