+++ src/Libs/GObject-2.0/Public/SignalTSender.cs
@@ -39,7 +39,7 @@ public class Signal<TSender> : SignalDefinition
             handle: sender.Handle
         );
 
-        sender.SignalConnectClosure(this, signalHandler, closure, after, detail);
+        sender.Handle.Connect(this, signalHandler, closure, after, detail);
     }
 
     /// <summary>
@@ -49,6 +49,6 @@ public class Signal<TSender> : SignalDefinition
     /// <param name="signalHandler">The signal handler function.</param>
     public void Disconnect(TSender sender, SignalHandler<TSender> signalHandler)
     {
-        sender.Disconnect(this, signalHandler);
+        sender.Handle.Disconnect(this, signalHandler);
     }
 }
